Remove live change to /tmp/ mount (see desc)
When doing these changes in place it causes the startup script behaviour
to fail. This is likely because the restart of waagent breaks the
back-end connection to the startup script.

Seeing as the process for building a node requires a reboot, this can be
removed without impacting how a node works in the live environment.